+++ containers_slice.go
@@ -1,4 +1,4 @@
-// Copyright 2017 Pilosa Corp.
+// Copyright (C) 2017-2018 Pilosa Corp. All rights reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -46,25 +46,19 @@ func (sc *sliceContainers) Put(key uint64, c *Container) {
 
 }
 
-func (sc *sliceContainers) PutContainerValues(key uint64, typ byte, n int, mapped bool) {
+func (sc *sliceContainers) PutContainerValues(key uint64, containerType byte, n int, mapped bool) {
 	i := search64(sc.keys, key)
 	if i < 0 {
 		c := NewContainer()
-		c.setTyp(typ)
-		c.setN(int32(n))
-		c.setMapped(mapped)
+		c.containerType = containerType
+		c.n = int32(n)
+		c.mapped = mapped
 		sc.insertAt(key, c, -i-1)
 	} else {
-		// if the container already exists, and is frozen, this may
-		// result in copying its data, which is sort of pointless
-		// because PutContainerValues almost always gets called
-		// because we're reading new data from a file -- but also
-		// that means this case probably never happens.
-		c := sc.containers[i].Thaw()
-		c.setTyp(typ)
-		c.setN(int32(n))
-		c.setMapped(mapped)
-		sc.containers[i] = c
+		c := sc.containers[i]
+		c.containerType = containerType
+		c.n = int32(n)
+		c.mapped = mapped
 	}
 
 }
@@ -74,10 +68,6 @@ func (sc *sliceContainers) Remove(key uint64) {
 	i := search64(sc.keys, key)
 	if i < 0 {
 		return
-	}
-	if key == sc.lastKey {
-		sc.lastKey = ^uint64(0)
-		sc.lastContainer = nil
 	}
 	sc.keys = append(sc.keys[:i], sc.keys[i+1:]...)
 	sc.containers = append(sc.containers[:i], sc.containers[i+1:]...)
@@ -124,17 +114,6 @@ func (sc *sliceContainers) Clone() Containers {
 	return other
 }
 
-func (sc *sliceContainers) Freeze() Containers {
-	other := newSliceContainers()
-	other.keys = make([]uint64, len(sc.keys))
-	other.containers = make([]*Container, len(sc.containers))
-	copy(other.keys, sc.keys)
-	for i, c := range sc.containers {
-		other.containers[i] = c.Freeze()
-	}
-	return other
-}
-
 func (sc *sliceContainers) Last() (key uint64, c *Container) {
 	if len(sc.keys) == 0 {
 		return 0, nil
@@ -150,7 +129,7 @@ func (sc *sliceContainers) Size() int {
 func (sc *sliceContainers) Count() uint64 {
 	n := uint64(0)
 	for i := range sc.containers {
-		n += uint64(sc.containers[i].N())
+		n += uint64(sc.containers[i].n)
 	}
 	return n
 }
@@ -158,18 +137,6 @@ func (sc *sliceContainers) Count() uint64 {
 func (sc *sliceContainers) Reset() {
 	sc.keys = sc.keys[:0]
 	sc.containers = sc.containers[:0]
-	sc.lastContainer = nil
-	sc.lastKey = 0
-}
-
-func (sc *sliceContainers) ResetN(n int) {
-	if cap(sc.keys) < n {
-		sc.keys = make([]uint64, 0, n)
-		sc.containers = make([]*Container, 0, n)
-	} else {
-		sc.keys = sc.keys[:0]
-		sc.containers = sc.containers[:0]
-	}
 	sc.lastContainer = nil
 	sc.lastKey = 0
 }
@@ -195,40 +162,6 @@ func (sc *sliceContainers) Repair() {
 	}
 }
 
-// Update calls fn (existing-container, existed), and expects
-// (new-container, write). If write is true, the container is used to
-// replace the given container.
-func (sc *sliceContainers) Update(key uint64, fn func(*Container, bool) (*Container, bool)) {
-	i, found := sc.seek(key)
-	var nc *Container
-	var write bool
-	if found {
-		nc, write = fn(sc.containers[i], true)
-		if write {
-			sc.containers[i] = nc
-		}
-	} else {
-		nc, write = fn(nil, false)
-		// don't expand the slice just to add a nil container, we
-		// could return that anyway
-		if write && nc != nil {
-			sc.insertAt(key, nc, -i-1)
-		}
-	}
-}
-
-// UpdateEvery calls fn (existing-container, existed), and expects
-// (new-container, write). If write is true, the container is used to
-// replace the given container.
-func (sc *sliceContainers) UpdateEvery(fn func(uint64, *Container, bool) (*Container, bool)) {
-	for i, c := range sc.containers {
-		nc, write := fn(sc.keys[i], c, true)
-		if write {
-			sc.containers[i] = nc
-		}
-	}
-}
-
 type sliceIterator struct {
 	e     *sliceContainers
 	i     int
@@ -237,20 +170,14 @@ type sliceIterator struct {
 }
 
 func (si *sliceIterator) Next() bool {
-	if si.e == nil {
+	if si.e == nil || si.i > len(si.e.keys)-1 {
 		return false
 	}
-	// discard nil containers from iteration. we don't always
-	// actually remove them because copying is expensive.
-	for si.i < len(si.e.keys) {
-		si.key = si.e.keys[si.i]
-		si.value = si.e.containers[si.i]
-		si.i++
-		if si.value != nil {
-			return true
-		}
-	}
-	return false
+	si.key = si.e.keys[si.i]
+	si.value = si.e.containers[si.i]
+	si.i++
+
+	return true
 }
 
 func (si *sliceIterator) Value() (uint64, *Container) {
